requirements.txt
anyascii==0.3.2
asgiref==3.8.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
cryptography==43.0.3
defusedxml==0.7.1
Django==4.2.16
django-appconf==1.0.6
django-cogwheels==0.3
django-compressor==4.5.1
django-crispy-forms==2.3
django-extensions==3.2.3
django-filter==24.3
django-modelcluster==6.3
django-permissionedforms==0.1
django-recaptcha==4.0.0
django-taggit==5.0.1
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
draftjs-exporter==5.0.0
et-xmlfile==1.1.0
Fabric3==1.14.post1
filetype==1.2.0
gunicorn==23.0.0
idna==3.8
l18n==2021.3
laces==0.1.1
lesscpy==0.15.1
openpyxl==3.1.5
packaging==24.2
paramiko==2.12.0
pillow==10.4.0
pillow_heif==0.18.0
ply==3.11
psycopg==3.2.1
psycopg2-binary==2.9.9
pycparser==2.22
PyNaCl==1.5.0
pytz==2024.1
rcssmin==1.1.2
requests==2.32.3
rjsmin==1.2.2
sentry-sdk==2.39.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
telepath==0.3.1
typing_extensions==4.12.2
urllib3==2.2.2
wagtail==6.2.1
wagtail-modeladmin==2.0.0
wagtail-seo==3.1.1
wagtailmenus==4.0.1
Willow==1.8.0
